+++ src/enums/generators/identifier-name-generators/IdentifierNameGenerator.ts
@@ -1,4 +1,0 @@
-export enum IdentifierNameGenerator {
-    HexadecimalIdentifierNameGenerator,
-    MangledIdentifierNameGenerator
-}