+++ src/enums/generators/identifier-names-generators/IdentifierNamesGenerator.ts
@@ -1,0 +1,4 @@
+export enum IdentifierNamesGenerator {
+    HexadecimalIdentifierNamesGenerator = 'hexadecimal',
+    MangledIdentifierNamesGenerator = 'mangled'
+}