+++ src/types/container/generators/TIdentifierNamesGeneratorFactory.d.ts
@@ -1,0 +1,4 @@
+import { IIdentifierNamesGenerator } from '../../../interfaces/generators/identifier-names-generators/IIdentifierNamesGenerator';
+import { IOptions } from '../../../interfaces/options/IOptions';
+
+export type TIdentifierNamesGeneratorFactory = (options: IOptions) => IIdentifierNamesGenerator;