+++ test/dev/dev.ts
@@ -35,7 +35,7 @@ import { NO_CUSTOM_NODES_PRESET } from '../../src/options/presets/NoCustomNodes'
         {
             ...NO_CUSTOM_NODES_PRESET,
             compact: false,
-            mangle: true
+            identifierNamesGenerator: 'mangled'
         }
     ).getObfuscatedCode();