+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Change Log
 ===
 v0.13.0
 ---
+* **Breaking change:** `mangle` option was removed.
+* **New option:** `identifierNamesGenerator` allows to set identifier names generator (`hexadecimal` or `mangled`).
 * **Breaking change:** all CLI options were renamed to `kebab-case` format (`--disableConsoleOutout` -> `--disable-console-output`)
 * Implemented custom `mangle` option algorithm without `esmangle`; fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/110